absract NowService
