+++ server/src/application.ts
@@ -1,5 +1,5 @@
 import { BootMixin } from '@loopback/boot';
-import { ApplicationConfig, BindingKey } from '@loopback/core';
+import { ApplicationConfig, BindingScope, CoreTags } from '@loopback/core';
 import { RestExplorerBindings, RestExplorerComponent } from '@loopback/rest-explorer';
 import { RestApplication } from '@loopback/rest';
 
@@ -7,27 +7,18 @@ import path from 'path';
 
 import { MainSequence } from './sequence';
 import { RadiodBindings } from './keys';
-import { NowService } from './services/now.service'
+import { NowSpotify } from './now/now.spotify'
 
 export class RadiodApplication extends BootMixin(RestApplication) {
 
   constructor(options: ApplicationConfig = {}) {
     super(options);
 
+    /* Loopback BINDINGS */
     this.projectRoot = __dirname;
-    this.bind(RadiodBindings.PROJECT_ROOT).to(path.join(__dirname, '../..'));
-
     this.sequence(MainSequence);
-
-    this.static('/', path.join(__dirname, '../../static'));
-    this.static('/jingles', path.join(__dirname, '../../static/jingles.html'));
-
     this.bind(RestExplorerBindings.CONFIG).to({ path: '/explorer' });
-
     this.component(RestExplorerComponent);
-
-    this.service(NowService);
-
     this.bootOptions = {
       controllers: {
         dirs: ['controllers'],
@@ -35,5 +26,17 @@ export class RadiodApplication extends BootMixin(RestApplication) {
         nested: true,
       },
     };
+
+    /* Static BINDINGS */
+    this.static('/', path.join(__dirname, '../../static'));
+    this.static('/jingles', path.join(__dirname, '../../static/jingles.html'));
+
+    /* Application BINDINGS */
+    this.bind(RadiodBindings.PROJECT_ROOT).to(path.join(__dirname, '../..'));
+
+    this.bind(RadiodBindings.NOW_SERVICE)
+      .toClass(NowSpotify)
+      .tag(CoreTags.LIFE_CYCLE_OBSERVER)
+      .inScope(BindingScope.SINGLETON);
   }
 }
